requirements.txt
# Core Django Framework
Django==5.2.5

# HTTP Requests for API calls
requests==2.31.0

# Configuration management for environment variables
python-decouple==3.8

# Google Generative AI (Gemini)
google-generativeai==0.8.3

# Image processing for stock logos
Pillow==10.1.0

# Development and Testing Dependencies
# (Uncomment for development)
# pytest==7.4.3
# pytest-django==4.7.0
# django-debug-toolbar==4.2.0

# Production Dependencies
# (Uncomment for production deployment)
# gunicorn==21.2.0
# whitenoise==6.6.0
# django-environ==0.11.2

# Security and Performance Enhancements
# (Optional but recommended)
# django-cors-headers==4.3.1
# django-ratelimit==4.1.0
